Collection override endpoints #17

+++ packages/tatsy-collection/index.js
@@ -5,26 +5,37 @@ module.exports = (name, options) => {
   const {
     schema = {},
     schemaOptions = {},
-    transform = (doc, obj) => obj
+    transform = (doc, obj) => obj,
+    endpoints = {}
   } = options;
 
   return (url, { Http, Mongo }) => {
-    const Model = Mongo.model(name, new Mongo.Schema(schema, {
-      timestamps: true,
-      versionKey: false,
-      toJSON: { transform },
-      ...schemaOptions
-    }));
+    const Model = Mongo.collection(name, schema, transform, schemaOptions);
 
     // getAll
-    Http.get(apiUrl(url), (req, res) => {
+    Http.get(apiUrl(url), async (req, res) => {
+      const handler = endpoints.getAll;
+
+      if (handler) {
+        return res.json(await handler.call({
+          collections: Mongo.Collections
+        }));
+      }
+
       Model.find({ }, (err, list) => {
         return res.json(returnJSON('success', list));
       });
     });
 
-    Http.get(apiUrl(url, '_id'), (req, res) => {
+    Http.get(apiUrl(url, '_id'), async (req, res) => {
+      const handler = endpoints.get;
       const { _id } = req.params;
+
+      if (handler) {
+        return res.json(await handler.call({
+          collections: Mongo.Collections,
+        }, _id));
+      }
 
       return Model.findOne({ _id }, (err, doc) => {
         if (doc) {
@@ -35,11 +46,18 @@ module.exports = (name, options) => {
       });
     });
 
-    Http.post(apiUrl(url), (req, res) => {
+    Http.post(apiUrl(url), async (req, res) => {
+      const handler = endpoints.post;
       const { body } = req;
 
       // create model
       const doc = new Model(body);
+
+      if (handler) {
+        return res.json(await handler.call({
+          collections: Mongo.Collections,
+        }, body));
+      }
 
       // save doc
       return doc.save((err, data) => {
@@ -47,9 +65,16 @@ module.exports = (name, options) => {
       });
     });
 
-    Http.put(apiUrl(url, '_id'), (req, res) => {
+    Http.put(apiUrl(url, '_id'), async (req, res) => {
+      const handler = endpoints.put;
       const { _id } = req.params;
       const { body } = req;
+
+      if (handler) {
+        return res.json(await handler.call({
+          collections: Mongo.Collections
+        }, _id, body));
+      }
 
       // update
       Model.findOneAndUpdate({ _id }, body, (err, data) => {
@@ -61,8 +86,15 @@ module.exports = (name, options) => {
       });
     });
 
-    Http.delete(apiUrl(url, '_id'), (req, res) => {
+    Http.delete(apiUrl(url, '_id'), async (req, res) => {
+      const handler = endpoints.delete;
       const { _id } = req.params;
+
+      if (handler) {
+        return res.json(await handler.call({
+          collections: Mongo.Collections,
+        }, _id));
+      }
 
       // remove
       return Model.findOneAndDelete({ _id }, (err, data) => {
@@ -73,5 +105,7 @@ module.exports = (name, options) => {
         return res.status(404).json(returnFail(_id));
       });
     });
+  
+    return Model;
   };
 };
